Update example with configuration in Readme
Currently, the reader may get an impression that `Configuration` can be changed dynamically, after `ImagePickerController` has been initialized. However, it does not support this feature, so I've changed example to reflect this case better.
I use `configuration` as a parameter to initialize `ImagePickerController`, so that it is clear, it cannot be mutated after initialization.